refactor: clean up file-naming
