+++ app/lib/components/Auth/Connection.js
@@ -1,11 +1,10 @@
 import React from 'react'
 import { H1, Paragraph } from '../typography/Typography'
 import { PrimaryButton, SecondaryButton } from '../elements/Button/Button'
-import { Wrap } from '../View/Wrapper'
+import { Wrap } from '../view/Wrapper'
 import { approveConnection } from '../../services/auth'
 
 const Connection = ({ connectionRequest, onApprove }) => {
-
   const onApproveConnection = async () => {
     await approveConnection(connectionRequest)
     onApprove()
@@ -28,40 +27,3 @@ const Connection = ({ connectionRequest, onApprove }) => {
 }
 
 export default Connection
-
-// export default class RegisterScreen extends React.Component {
-//   state = {
-//     firstName: '',
-//     lastName: '',
-//   }
-
-//   handleSubmit = () => {
-//     const { firstName, lastName } = this.state
-
-//     if (firstName !== '' && lastName !== '') {
-//       this.props.onSubmit(this.state)
-//     }
-//   }
-
-//   render() {
-//     return (
-//       <>
-//         <H1>MyData</H1>
-//         <Input
-//           name="firstName"
-//           onChangeText={firstName => this.setState({ firstName })}
-//           placeholder="Förnamn"
-//           style={{ marginTop: 64 }}
-//           value={this.state.firstName}
-//         />
-//         <Input
-//           name="lastName"
-//           onChangeText={lastName => this.setState({ lastName })}
-//           placeholder="Efternamn"
-//           value={this.state.lastName}
-//         />
-//         <PrimaryButton onPress={this.handleSubmit}>Registrera</PrimaryButton>
-//       </>
-//     )
-//   }
-// }
