+++ app/lib/components/wizard/Finished.js
