+++ app/lib/components/wizard/img/finished.png
