+++ app/lib/components/wizard/img/storage.png
