+++ app/lib/screens/AuthLoading/AuthLoading.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { StatusBar } from 'react-native'
-import { Wrap } from '../../components/View/Wrapper'
+import { Wrap } from '../../components/view/Wrapper'
 import { Spinner } from '../../components/elements/Spinner/Spinner'
 import { getAccount } from '../../services/account'
