+++ app/lib/screens/Home/Home.js
@@ -2,7 +2,7 @@
 import React, { useState, useEffect } from 'react'
 import { Linking } from 'react-native'
 import { H1 } from '../../components/typography/Typography'
-import { Wrap, ScrollViewWrap } from '../../components/View/Wrapper'
+import { Wrap, ScrollViewWrap } from '../../components/view/Wrapper'
 import ConsentList from '../../components/ConsentList/ConsentList'
 import { getAll } from '../../services/consents'
 import { getAccount } from '../../services/account'
@@ -19,6 +19,7 @@ const HomeScreen = ({ navigation }) => {
 
   useEffect(() => {
     Linking.addEventListener('url', handleOpenURL)
+
     getAccount().then(console.log)
     // TODO: Uncomment =)
     // const subscription = navigation.addListener('willFocus', updateConsents)
@@ -28,7 +29,7 @@ const HomeScreen = ({ navigation }) => {
       Linking.removeEventListener('url', handleOpenURL)
       // subscription.remove()
     }
-  }, [])
+  }, [navigation])
 
   return (
     <Wrap style={{ justifyContent: 'flex-start', paddingHorizontal: 26 }}>
